Added Pomodoro options to Jam countdown
Under Project > Tools there are a series of options which automatically assign the "end time" for the "jam counter" to a specified number of minutes after the current time. It does some time-zone calculations to make sure this works in local time as the addon was made for local-time calculations.

There is a known edge-case with OS.get_time_zone_info where on Windows machines during daylight savings hours, the returned bias is incorrect. This is a currently existing issue with the Godot engine, it is assumed when that issue is resolved, this edge case will be removed too.
Current recommended solution: Add a value to the UNIX_TIME_CALC_CORRECTION in the plugin during daylight savings time. There should be enough comments around the calculations to make it make sense